+++ extensions/ext-llm-anthropic/src/anthropic-request-builder.test.ts
@@ -1,0 +1,217 @@
+import { assertEquals } from "#veryfront/testing/assert.ts";
+import { describe, it } from "#veryfront/testing/bdd.ts";
+import type { RuntimePromptMessage } from "veryfront/provider/shared";
+import { buildAnthropicMessagesRequest } from "./anthropic-request-builder.ts";
+
+function createWarningCollector() {
+  const warnings: Array<{
+    type: "unsupported-setting" | "other";
+    setting?: string;
+    details?: string;
+    provider: string;
+  }> = [];
+
+  return {
+    push(warning: {
+      type: "unsupported-setting" | "other";
+      setting?: string;
+      details?: string;
+      provider: string;
+    }) {
+      warnings.push(warning);
+    },
+    drain() {
+      return warnings.slice();
+    },
+  };
+}
+
+describe("ext-llm-anthropic/anthropic-request-builder", () => {
+  it("preserves Messages request shaping, provider option merge order, and warnings", () => {
+    const prompt: RuntimePromptMessage[] = [
+      { role: "system", content: "You are careful." },
+      {
+        role: "user",
+        content: [
+          { type: "text", text: "Inspect this." },
+          { type: "image", mediaType: "image/png", url: "https://example.test/image.png" },
+        ],
+      },
+      {
+        role: "assistant",
+        content: [
+          { type: "text", text: "I will check." },
+          {
+            type: "tool-call",
+            toolCallId: "tool_1",
+            toolName: "lookup",
+            input: { id: "abc" },
+          },
+          {
+            type: "reasoning",
+            text: "Thinking trace",
+            signature: "sig_123",
+          },
+        ],
+      },
+      {
+        role: "tool",
+        content: [{
+          type: "tool-result",
+          toolCallId: "tool_1",
+          toolName: "lookup",
+          output: { type: "json", value: { ok: true } },
+        }],
+      },
+    ];
+    const warnings = createWarningCollector();
+
+    const body = buildAnthropicMessagesRequest(
+      "claude-sonnet-4-5-20250929",
+      "bedrock",
+      {
+        prompt,
+        maxOutputTokens: 20_000,
+        temperature: 0.4,
+        topP: 0.9,
+        topK: 10,
+        stopSequences: ["one", "two", "three", "four", "five"],
+        tools: [
+          {
+            type: "function",
+            name: "lookup",
+            description: "Look up a value",
+            inputSchema: {
+              jsonSchema: { type: "object", properties: { id: { type: "string" } } },
+            },
+          },
+          {
+            type: "provider",
+            name: "web",
+            id: "anthropic.web_search",
+            args: { maxUses: 2 },
+          },
+        ],
+        toolChoice: "auto",
+        seed: 7,
+        presencePenalty: 0.1,
+        frequencyPenalty: 0.2,
+        cacheControl: { system: true, tools: "1h" },
+        reasoning: { enabled: true, effort: "high" },
+        responseFormat: { type: "json" },
+        userId: "user_123",
+        mcpServers: [{
+          type: "url",
+          url: "https://example.test/mcp",
+          authorizationToken: "token_123",
+          toolConfiguration: {
+            allowedTools: ["read_file"],
+          },
+        }],
+        anthropicContainer: { id: "ctr_1" },
+        providerOptions: {
+          anthropic: {
+            custom_anthropic: true,
+            max_tokens: 222,
+          },
+          bedrock: {
+            custom_bedrock: true,
+            temperature: 0.1,
+          },
+        },
+      },
+      true,
+      warnings,
+    );
+
+    assertEquals(body, {
+      model: "claude-sonnet-4-5-20250929",
+      messages: [
+        {
+          role: "user",
+          content: [
+            { type: "text", text: "Inspect this." },
+            {
+              type: "image",
+              source: {
+                type: "url",
+                url: "https://example.test/image.png",
+              },
+            },
+          ],
+        },
+        {
+          role: "assistant",
+          content: [
+            { type: "text", text: "I will check." },
+            {
+              type: "tool_use",
+              id: "tool_1",
+              name: "lookup",
+              input: { id: "abc" },
+            },
+            {
+              type: "thinking",
+              thinking: "Thinking trace",
+              signature: "sig_123",
+            },
+          ],
+        },
+        {
+          role: "user",
+          content: [{
+            type: "tool_result",
+            tool_use_id: "tool_1",
+            content: '{"ok":true}',
+          }],
+        },
+      ],
+      max_tokens: 222,
+      stream: true,
+      system: [{
+        type: "text",
+        text: "You are careful.",
+        cache_control: { type: "ephemeral" },
+      }],
+      stop_sequences: ["one", "two", "three", "four"],
+      tools: [
+        {
+          name: "lookup",
+          description: "Look up a value",
+          input_schema: { type: "object", properties: { id: { type: "string" } } },
+        },
+        {
+          type: "web_search_20250305",
+          name: "web",
+          max_uses: 2,
+          cache_control: { type: "ephemeral", ttl: "1h" },
+        },
+      ],
+      tool_choice: { type: "auto" },
+      thinking: { type: "enabled", budget_tokens: 16_384 },
+      metadata: { user_id: "user_123" },
+      mcp_servers: [{
+        type: "url",
+        url: "https://example.test/mcp",
+        authorization_token: "token_123",
+        tool_configuration: {
+          allowed_tools: ["read_file"],
+        },
+      }],
+      container: { id: "ctr_1" },
+      custom_anthropic: true,
+      custom_bedrock: true,
+      temperature: 0.1,
+    });
+    assertEquals(warnings.drain().map((warning) => warning.setting), [
+      "presencePenalty",
+      "frequencyPenalty",
+      "seed",
+      "topK",
+      "stopSequences",
+      "temperature",
+      "topP",
+      "responseFormat",
+    ]);
+  });
+});
